% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{add_weekly_variation}
\alias{add_weekly_variation}
\title{Return weekly variation of an asset}
\usage{
add_weekly_variation(stock_data, start_date)
}
\arguments{
\item{stock_data}{data.frame}

\item{start_date}{character}
}
\value{
data.frame
}
\description{
Given a \code{stock_data} data.frame, increment data with a \code{Weekly_Variation} column.
}
\examples{
raw_stock = stock("VALE3.SA")
add_weekly_variation(raw_stock, '2019-10-26')
}
