% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta}
\alias{beta}
\title{Calculate beta index for a given asset in the last 12 months.}
\usage{
beta(stock_code)
}
\arguments{
\item{stock_code}{character}
}
\value{
character
}
\description{
Calculate beta index for a given asset in the last 12 months.
}
\examples{
beta("VALE3.SA")
}
