% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Display a graph of an asset variation over months}
\usage{
graph(stock_code, start_date)
}
\arguments{
\item{stock_code}{character}

\item{start_date}{character}
}
\value{
ggplot object
}
\description{
Display a graph of an asset variation over months
}
\examples{
graph("VALE3.SA", "2019-10-31")
}
