% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint.R
\name{hint}
\alias{hint}
\title{Hint action for a stock}
\usage{
hint(stock_code)
}
\arguments{
\item{stock_code}{character}
}
\value{
character
}
\description{
Suggests a action (buy, hold or sell) for a specific stock.
}
\examples{
hint("VALE3.SA")
}
