% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{stock}
\alias{stock}
\title{Return stock data}
\usage{
stock(stock_code)
}
\arguments{
\item{stock_code}{character}
}
\value{
data.frame
}
\description{
Given a \code{stock_code}, return its price history from Yahoo Finance.
}
\examples{
stock("VALE3.SA")
}
