% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{weekly_graph}
\alias{weekly_graph}
\title{Display a graph of an asset variation over weeks}
\usage{
weekly_graph(stock_code, start_date)
}
\arguments{
\item{stock_code}{character}

\item{start_date}{character}
}
\value{
ggplot object
}
\description{
Display a graph of an asset variation over weeks
}
\examples{
weekly_graph("VALE3.SA", "2019-10-31")
}
