% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{add_monthly_variation}
\alias{add_monthly_variation}
\title{Return monthly variation of a asset}
\usage{
add_monthly_variation(stock_data, start_date)
}
\arguments{
\item{stock_data}{data.frame}
}
\value{
data.frame
}
\description{
Given a \code{stock_data} data.frame, increment data with a \code{Variation}.
}
\examples{
raw_stock = stock("VALE3.SA")
add_monthly_variation(raw_stock, '2019-10-26')
}
